requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
sentence-transformers
streamlit
openai
python-dotenv
requests
